# docs/swagger.yaml
basePath: /
definitions:
...
  delete.MessageResponse:
    properties:
      message:
        type: string
    type: object
  getbyid.ResponseError:
    properties:
      error:
        type: string
    type: object
...
  ukk-smkn2_features_fileupload_create.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ukk-smkn2_features_fileupload_create.Request:
    properties:
      karya_id:
        type: string
      nama:
        type: string
      tipe:
        type: string
      url:
        type: string
    type: object
  ukk-smkn2_features_fileupload_create.Response:
    properties:
      file_id:
        type: string
      karya_id:
        type: string
      nama:
        type: string
      tipe:
        type: string
      url:
        type: string
    type: object
  ukk-smkn2_features_fileupload_delete.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ukk-smkn2_features_fileupload_delete.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  ukk-smkn2_features_fileupload_getall.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ukk-smkn2_features_fileupload_getall.Response:
    properties:
      file_id:
        type: string
      karya_id:
        type: string
      nama:
        type: string
      tipe:
        type: string
      url:
        type: string
    type: object
  ukk-smkn2_features_fileupload_getbyid.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ukk-smkn2_features_fileupload_getbyid.Response:
    properties:
      file_id:
        type: string
      karya_id:
        type: string
      nama:
        type: string
      tipe:
        type: string
      url:
        type: string
    type: object
  ukk-smkn2_features_fileupload_update.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ukk-smkn2_features_fileupload_update.Request:
    properties:
      karya_id:
        type: string
      nama:
        type: string
      tipe:
        type: string
      url:
        type: string
    type: object
  ukk-smkn2_features_fileupload_update.Response:
    properties:
      file_id:
        type: string
      karya_id:
        type: string
      nama:
        type: string
      tipe:
        type: string
      url:
        type: string
    type: object
  ukk-smkn2_features_karyaukk_create.ErrorResponse:
    properties:
      error:
        example: Deskripsi error
        type: string
    type: object
  ukk-smkn2_features_karyaukk_create.Request:
    properties:
      deskripsi:
        type: string
      judul:
        type: string
      kategori_id:
        type: string
      link:
        type: string
      siswa_id:
        type: string
    type: object
  ukk-smkn2_features_karyaukk_create.Response:
    properties:
      deskripsi:
        type: string
      judul:
        type: string
      karya_id:
        type: string
      kategori_id:
        type: string
      link:
        type: string
      siswa_id:
        type: string
    type: object
  ukk-smkn2_features_karyaukk_delete.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ukk-smkn2_features_karyaukk_getall.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ukk-smkn2_features_karyaukk_getall.Response:
    properties:
      created_at:
        type: string
      deskripsi:
        type: string
      judul:
        type: string
      karya_id:
        type: string
      kategori_id:
        type: string
      link:
        type: string
      siswa_id:
        type: string
      updated_at:
        type: string
    type: object
  ukk-smkn2_features_karyaukk_getbyid.Response:
    properties:
      created_at:
        type: string
      deskripsi:
        type: string
      judul:
        type: string
      karya_id:
        type: string
      kategori_id:
        type: string
      link:
        type: string
      siswa_id:
        type: string
      updated_at:
        type: string
    type: object
  ukk-smkn2_features_karyaukk_update.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ukk-smkn2_features_karyaukk_update.Request:
    properties:
      deskripsi:
        type: string
      judul:
        type: string
      kategori_id:
        type: string
      link:
        type: string
      siswa_id:
        type: string
    type: object
  ukk-smkn2_features_karyaukk_update.Response:
    properties:
      deskripsi:
        type: string
      judul:
        type: string
      karya_id:
        type: string
      kategori_id:
        type: string
      link:
        type: string
      siswa_id:
        type: string
    type: object
...
host: 127.0.0.1:8080
info:
  contact: {}
  description: Restful API using go fiber
  title: your own API application
  version: "1.0"
paths:
...
  /api/fileupload:
    get:
      description: Mengambil semua data FileUpload
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ukk-smkn2_features_fileupload_getall.Response'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ukk-smkn2_features_fileupload_getall.ErrorResponse'
      summary: Get All FileUploads
      tags:
      - FileUpload
    post:
      consumes:
      - application/json
      description: Membuat data file upload baru
      parameters:
      - description: Data file upload
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/ukk-smkn2_features_fileupload_create.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ukk-smkn2_features_fileupload_create.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ukk-smkn2_features_fileupload_create.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ukk-smkn2_features_fileupload_create.ErrorResponse'
      summary: Create FileUpload
      tags:
      - FileUpload
  /api/fileupload/{id}:
    delete:
      description: Menghapus data FileUpload berdasarkan ID
      parameters:
      - description: FileUpload ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ukk-smkn2_features_fileupload_delete.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ukk-smkn2_features_fileupload_delete.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ukk-smkn2_features_fileupload_delete.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ukk-smkn2_features_fileupload_delete.ErrorResponse'
      summary: Delete FileUpload by ID
      tags:
      - FileUpload
    get:
      description: Mengambil data FileUpload berdasarkan ID
      parameters:
      - description: FileUpload ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ukk-smkn2_features_fileupload_getbyid.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ukk-smkn2_features_fileupload_getbyid.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ukk-smkn2_features_fileupload_getbyid.ErrorResponse'
      summary: Get FileUpload by ID
      tags:
      - FileUpload
    put:
      consumes:
      - application/json
      description: Mengupdate data FileUpload berdasarkan ID
      parameters:
      - description: FileUpload ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Data file yang diupdate
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ukk-smkn2_features_fileupload_update.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ukk-smkn2_features_fileupload_update.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ukk-smkn2_features_fileupload_update.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ukk-smkn2_features_fileupload_update.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ukk-smkn2_features_fileupload_update.ErrorResponse'
      summary: Update FileUpload by ID
      tags:
      - FileUpload
  /api/karya-ukk/{id}:
    get:
      consumes:
      - application/json
      description: Mengambil data Karya UKK berdasarkan ID
      parameters:
      - description: Karya UKK ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ukk-smkn2_features_karyaukk_getbyid.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/getbyid.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/getbyid.ResponseError'
      security:
      - BearerAuth: []
      summary: Get Karya UKK by ID
      tags:
      - KaryaUKK
  /api/karyaukk:
    get:
      description: Mengambil semua data karya UKK
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ukk-smkn2_features_karyaukk_getall.Response'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ukk-smkn2_features_karyaukk_getall.ErrorResponse'
      summary: Get all KaryaUKK
      tags:
      - KaryaUKK
    post:
      consumes:
      - application/json
      description: Membuat karya UKK baru dengan judul, deskripsi, link, siswa, dan
        kategori
      parameters:
      - description: Request body untuk membuat karya UKK
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ukk-smkn2_features_karyaukk_create.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ukk-smkn2_features_karyaukk_create.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ukk-smkn2_features_karyaukk_create.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ukk-smkn2_features_karyaukk_create.ErrorResponse'
      summary: Create a new KaryaUKK
      tags:
      - KaryaUKK
  /api/karyaukk/{id}:
    delete:
      description: Menghapus karya UKK berdasarkan ID yang diberikan
      parameters:
      - description: KaryaUKK ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/delete.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ukk-smkn2_features_karyaukk_delete.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ukk-smkn2_features_karyaukk_delete.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ukk-smkn2_features_karyaukk_delete.ErrorResponse'
      summary: Delete a KaryaUKK by ID
      tags:
      - KaryaUKK
    put:
      consumes:
      - application/json
      description: Mengupdate karya UKK berdasarkan ID yang diberikan
      parameters:
      - description: KaryaUKK ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Data karya UKK yang akan diupdate
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/ukk-smkn2_features_karyaukk_update.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ukk-smkn2_features_karyaukk_update.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ukk-smkn2_features_karyaukk_update.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ukk-smkn2_features_karyaukk_update.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ukk-smkn2_features_karyaukk_update.ErrorResponse'
      summary: Update a KaryaUKK by ID
      tags:
      - KaryaUKK
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"